Move contents of HcalDbProducer header to source file
